# api-gateway/src/test/resources/openapi/growbusiness-gateway-api.yaml
openapi: 3.0.3
info:
  title: GrowBusiness API Gateway
  description: |
    Semua endpoint microservices RSU Sumatera melalui **API Gateway** (`http://localhost:8000`).

    **Alur testing:**
    1. `POST /api/auth/login` dengan JSON (contoh admin di bawah)
    2. Salin `token` dari response
    3. Klik **Authorize** → masukkan `Bearer <token>`
    4. Panggil endpoint lain dengan body JSON sesuai contoh

    **Role:** ADMIN, DOCTOR, PHARMACIST, STAFF, USER
  version: 1.0.0
  contact:
    name: GrowBusiness RS
servers:
  - url: http://localhost:8000
    description: API Gateway (local)
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              admin:
                summary: Register Admin
                value:
                  username: admin
                  password: admin123
                  nama: Admin User
                  email: admin@test.com
                  role: ADMIN
              doctor:
                summary: Register Doctor
                value:
                  username: dokter1
                  password: pass123
                  nama: Dr. Budi
                  email: budi@test.com
                  role: DOCTOR
              pharmacist:
                summary: Register Pharmacist
                value:
                  username: apt1
                  password: pass123
                  nama: Apt. Siti
                  email: apt@test.com
                  role: PHARMACIST
              staff:
                summary: Register Staff
                value:
                  username: staff1
                  password: pass123
                  nama: Staff
                  email: staff@test.com
                  role: STAFF
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User registered successfully
                username: admin
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login (dapatkan JWT)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: admin
              password: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
              example:
                token: eyJhbGciOiJIUzUxMiJ9.example.token
                username: admin
                message: Login successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              username: admin
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/profile:
    get:
      tags: [Authentication]
      summary: Get user profile
      description: Returns current logged-in user information including ID. Requires Bearer Token.
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/medical/examinations:
    post:
      tags: [Medical Examinations]
      summary: Create medical examination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PemeriksaanRequest'
            example:
              pasienId: 1
              tekananDarah: 120/80
              beratBadan: 70.5
              tinggiBadan: 175.0
              suhuTubuh: 36.5
              keluhan: Sakit kepala ringan
              hasilPemeriksaan: Hasil pemeriksaan normal
              dokterId: DOK001
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponsePemeriksaan'
  /api/medical/examinations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Medical Examinations]
      summary: Get examination by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponsePemeriksaan'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Medical Examinations]
      summary: Update examination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PemeriksaanRequest'
            example:
              pasienId: 1
              tekananDarah: 125/85
              beratBadan: 71.0
              tinggiBadan: 175.0
              suhuTubuh: 36.8
              keluhan: Sakit kepala ringan, agak membaik
              hasilPemeriksaan: Hasil pemeriksaan membaik
              dokterId: DOK001
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponsePemeriksaan'
  /api/medical/examinations/patient/{pasienId}:
    get:
      tags: [Medical Examinations]
      summary: Get examinations for patient
      parameters:
        - name: pasienId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponsePemeriksaan'
  /api/medical/prescriptions:
    post:
      tags: [Prescriptions]
      summary: Create prescription (Kafka saga)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResepRequest'
            example:
              pemeriksaanId: 1
              pasienId: 1
              namaObat: Paracetamol
              dosis: 500mg
              frekuensi: 3x sehari
              jumlah: 10
              catatan: Minum setelah makan
              dokterNama: Dr. Budi
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseResep'
  /api/medical/prescriptions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Prescriptions]
      summary: Get prescription by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseResep'
  /api/medical/prescriptions/patient/{pasienId}:
    get:
      tags: [Prescriptions]
      summary: Get prescriptions for patient
      parameters:
        - name: pasienId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseResep'
...
  /api/medical/prescriptions/{id}/status:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: status
        in: query
        required: true
        schema:
          type: string
          enum: [PENDING, RESERVED, COMPLETED, CANCELLED, FAILED_OUT_OF_STOCK]
        example: COMPLETED
    patch:
      tags: [Prescriptions]
      summary: Update prescription status
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseResep'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token dari POST /api/auth/login. Format Authorize: Bearer <token>'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: admin123
    RegisterRequest:
      type: object
      required: [username, password, nama, email]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        nama:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, DOCTOR, PHARMACIST, STAFF, USER]
          default: USER
    RefreshTokenRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
          example: admin
    AuthTokenResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        message:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        username:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    UserProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin
        nama:
          type: string
          example: Admin User
        email:
          type: string
          format: email
          example: admin@test.com
        role:
          type: string
          example: ADMIN
        isActive:
          type: boolean
          example: true
...
    PemeriksaanRequest:
      type: object
      required: [pasienId]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        pasienId:
          type: integer
          format: int64
        tekananDarah:
          type: string
          example: 120/80
        beratBadan:
          type: number
          format: double
        tinggiBadan:
          type: number
          format: double
        suhuTubuh:
          type: number
          format: double
        keluhan:
          type: string
        hasilPemeriksaan:
          type: string
        dokterId:
          type: string
    ResepRequest:
      type: object
      required: [pemeriksaanId, pasienId, namaObat]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        pemeriksaanId:
          type: integer
          format: int64
        pasienId:
          type: integer
          format: int64
        namaObat:
          type: string
        dosis:
          type: string
        frekuensi:
          type: string
        jumlah:
          type: integer
        catatan:
          type: string
        status:
          type: string
          readOnly: true
        dokterNama:
          type: string
...
    DataResponsePemeriksaan:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/PemeriksaanRequest'
    ListResponsePemeriksaan:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PemeriksaanRequest'
        count:
          type: integer
    DataResponseResep:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/ResepRequest'
    ListResponseResep:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ResepRequest'
        count:
          type: integer
...